Allow user to choose DIGEST or BASIC for WebDAV authentication. This is critical for institutions with policies preventing use of multiple passwords, and to simplify authentication for user (only need one password).
This abviously necessitates https, but that's just good practice anyway.

The default remains DIGEST.

Notes:

The ldap driver change is from a previous personal commit of returning true
for userExists if there is existing storage for the user. It doesn't have
anything to do with the auth selection.

class.AJXP_Sabre_AuthBackend.php was simply renamed to reflect DIGEST.

In standard_conf_actions.xml, I reused a message about successfully activating
WebDAV, but it really should have a custom message because the existing
message reminds user to set a password.