boot.conf: Proposed improvement on user guidance
In first instance admins should be made aware of what makes
the difference in using a DBMS as config storage or a serialized
file.

At this stage they are not/less interested in knowing what type of DBMS,
only when they want to go with a DBMS, let them chose from the list of
DBMS supported by Pydio.