auth.ftp: Try to improve meaning
Actually the french translation looks much closer to the actual use of
of the auth.ftp module. It only verifies credentials against an FTP server
but doesn't actually open workspaces/repositories.
(well, actually only for testing the connection)

The Portueguese translation looks as if it was translated from the
english one. I can't fix it without being familiar with Portuguese.