auth.ftp: Allign english with french translation
Looking at the french translation and the user interface reveals
that this doesn't eventually have anything to do with testing the
connection but actually being the workspace used to validate credentials.

Since the translation in english was misleading I tried an automatic
translation of the Portoguese string as well, but is is far from
optimal.