missing "," character in various en.php files
For translators to allow correct transfer of the text .php files to the .po format,
so that a CAT tool like OmegaT can be used to ease the translation, text lines
need to end with ",". Currently, some of the php files were missing this
character on the last text line of the file. This commit fixes this.

list of files was generated with this command
for i in `find . -type f -name en.php`; do pcregrep -M '\"\n\)' "$i" && echo $i; done

files for other languages than EN were not touched by me, but can be listed with this:
for i in `find . -type f -name *.php`; do pcregrep -M '\"\n\)' "$i" && echo && echo $i; done | grep /i18n/