Extend buffer read to account for MSOOXML files
The original buffer read of 100 bytes is not enough to determine whether a application/zip file is actually a Microsoft Open Office XML file.  By reading more of the file, finfo can make a determination of the file type.  One of the issues is that the header size that required is dependent on the first archive file size.  I think 2KB will be enough to determine MSOOXML files, 3KB may be safer.