Fix Sort By when using pagination
Hi we just wrote that code to fix the Sort By , when using the pagination. 
The error code that we have was ".foreach is not a function" and then blank page. This code is using a FOR loop and drain the object inside the object properly.
Please push that code